fetch learns

+++ ddpg_her/replay_buffer.py
@@ -17,7 +17,7 @@ class ReplayBuffer:
         experiences = random.sample(self.memory, batch_size)
 
         states = torch.from_numpy(np.vstack([e.state for e in experiences if e is not None])).float()
-        actions = torch.from_numpy(np.vstack([e.action for e in experiences if e is not None])).long()
+        actions = torch.from_numpy(np.vstack([e.action for e in experiences if e is not None])).float()
         rewards = torch.from_numpy(np.vstack([e.reward for e in experiences if e is not None])).float()
         next_states = torch.from_numpy(np.vstack([e.next_state for e in experiences if e is not None])).float()
         dones = torch.from_numpy(np.vstack([e.done for e in experiences if e is not None]).astype(np.uint8)).float()
